backend/requirements.txt
Django==5.0.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.1.1
gunicorn==23.0.0
python-dotenv==1.0.1
Pillow==10.4.0
whitenoise==6.7.0

# Dev / testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
ruff==0.6.9
